ai_isp_simulator/requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
PyYAML>=6.0

# AI/ML frameworks
torch>=1.9.0
torchvision>=0.10.0
tensorflow>=2.6.0
onnx>=1.10.0
onnxruntime>=1.8.0

# Image processing
scikit-image>=0.18.0
scipy>=1.7.0

# Quality assessment
pytorch-fid>=0.2.0
lpips>=0.1.4
brisque>=0.1.0

# GUI and visualization
matplotlib>=3.4.0
seaborn>=0.11.0
streamlit>=1.0.0
PyQt5>=5.15.0

# Data handling
h5py>=3.3.0
pandas>=1.3.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Development
black>=21.0.0
flake8>=3.9.0
mypy>=0.910
